Add layout for single item tooltip
